# [shape_poly] Performance improvements for symbolic dimension manipulations (step 2)